server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   -r requirements.in
    #   voyageai
aiolimiter==1.2.1
    # via voyageai
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   typer-slim
    #   uvicorn
dnspython==2.8.0
    # via
    #   -r requirements.in
    #   email-validator
    #   pymongo
email-validator==2.3.0
    # via
    #   -r requirements.in
    #   pydantic
fastapi==0.120.2
    # via -r requirements.in
fastapi-cli==0.0.14
    # via -r requirements.in
fastapi-cloud-cli==0.3.1
    # via -r requirements.in
filelock==3.20.3
    # via
    #   -r requirements.in
    #   huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via huggingface-hub
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   fastapi-cloud-cli
    #   huggingface-hub
    #   langsmith
huggingface-hub==1.0.1
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
iniconfig==2.3.0
    # via pytest
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==1.2.5
    # via langchain-text-splitters
langchain-text-splitters==1.0.0
    # via voyageai
langsmith==0.4.40
    # via langchain-core
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numpy==2.3.4
    # via voyageai
orjson==3.11.5
    # via
    #   -r requirements.in
    #   langsmith
packaging==25.0
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   pytest
pillow==12.0.0
    # via voyageai
pluggy==1.6.0
    # via pytest
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic[email]==2.12.3
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   langchain-core
    #   langsmith
    #   voyageai
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pymongo==4.16.0
    # via -r requirements.in
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   langsmith
    #   requests-toolbelt
    #   voyageai
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via
    #   -r requirements.in
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   -r requirements.in
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.1
    # via fastapi-cloud-cli
sentry-sdk==2.42.1
    # via
    #   -r requirements.in
    #   fastapi-cloud-cli
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
sniffio==1.3.1
    # via anyio
starlette==0.49.1
    # via
    #   -r requirements.in
    #   fastapi
tenacity==9.1.2
    # via
    #   langchain-core
    #   voyageai
tokenizers==0.22.1
    # via voyageai
tqdm==4.67.1
    # via huggingface-hub
typer==0.20.0
    # via
    #   -r requirements.in
    #   fastapi-cli
    #   fastapi-cloud-cli
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   -r requirements.in
    #   requests
    #   sentry-sdk
uuid-utils==0.12.0
    # via langchain-core
uvicorn[standard]==0.38.0
    # via
    #   -r requirements.in
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via
    #   -r requirements.in
    #   uvicorn
voyageai==0.3.5
    # via -r requirements.in
watchfiles==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   uvicorn
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
